+++ algorithms/lesson-03-search/src/main/java/gu/geekbrains/search/ExponentialSearch.java
@@ -1,4 +1,4 @@
-package search;
+package gu.geekbrains.search;
 
 import java.util.function.Function;
