+++ algorithms/lesson-03-search/src/main/java/gu/geekbrains/search/InterpolationSearch.java
@@ -1,9 +1,9 @@
-package search;
+package gu.geekbrains.search;
 
 import java.util.List;
 import java.util.function.Function;
 
-public class BinarySearch {
+public class InterpolationSearch {
     public static <T extends Comparable<T>> int findIndex(List<T> list, T element) {
         return findIndex(list::get, list.size() - 1, element);
     }
@@ -14,7 +14,10 @@ public class BinarySearch {
 
     public static <T extends Comparable<T>> int findIndex(Function<Integer, T> calculate, int start, int end, T element) {
         while (start <= end) {
-            int i = (start + end) / 2;
+            int startHashCode = calculate.apply(start).hashCode();
+            int endHashCode = calculate.apply(end).hashCode();
+            double k = (double) (element.hashCode() - startHashCode) / (endHashCode - startHashCode);
+            int i = (int) ((end - start) * k + start);
 
             if (element.compareTo(calculate.apply(i)) == 0) {
                 return i;
